Remove of code that ping nylas servers on IMAP signup
When Nylas (the company) dropped supported for Nylas Mail and released
a version more suited for an open-source project they removed a number
of calls to the Nylas cloud API that was used to sync email. However,
they left in a call to the API that was triggered on signing up with
an IMAP/SMTP account. This call sends the details of the SMTP
authentication to Nylas's servers. The servers then, at a minimum,
attempt a connection to the SMTP server and return an error if that
fails. If it succeeds, the API returns an API token for Nylas Mail to
use with future connections to the API.

This commit removes the API call, as it's no longer useful and in fact
gets in the way when attempting to connect to a local SMTP server. It
also removes code that stores the API token, and some code that checks
that the API token exists. An API call with the same data to a *local*
API is left.