Remove delta connections code
This is only used for syncing with the N1CloudAPI, which is the API
interface to Nylas's cloud servers.  Nylas removed support for syncing
email with their cloud servers in a September code release aimed at
making it easier to run as a community project.

This code is no longer being used anywhere besides the developer bar,
which is monitor for now non-existant DeltaConnection for use when
debugging the application.