[client-sync] fix extracting multipart/alternative parts (#90)
When extracting multipart/alternative part, it is currently assumed that the mail contains at least one text/html part. While that's the case most of the time, it doesn't have to be always true. There are cases when multipart message contains just a single single text/plain part.

This fix conforms to RFC 2046:
"The use of the "multipart" media type with only a single body part may be useful in certain contexts, and is explicitly permitted."
"In general, the best choice is the LAST part of a type supported by the recipient system's local environment."