fix encoding and decoding of attachment filenames (#92)
* fix encoding and decoding filenames with special characters

* fallback to name in content-type when filename not present in disposition

removed filename encoding, as it's something that nodemailer does automatically and it resulted in filename being encoded twice when sending attachments. Fixes #91.

When decoding filename of received attachment, we should use parseMimeWords function instead of decodeMimeWord, because longer names can be composed of multiple encoded-words. Fixes #41.

When filename is not present in Content-Disposition header, we can use name in Content-Type as a fallback. Fixes #10.

updated nodemailer to the latest version. There should not be any BC breaks except adjusted oauth smtp config.