Display Contacts before Threads when searching
Its a more usable order to search for people (natural tendency than subject or message content).  Therefore it makes more sense to reverse the order in which these results are displayed in the Quick Search Results pop-over.  Also, when searching for a contact, the pop-over lists all the threads from the contact before displaying the contact!  This commit reverses the order and puts people first!