Fix url parsing (#149)
* Support to show ICS invite as attachments

ICS invites are currently not displayed correctly in Nylas.  This commit adds support to extract calendar files and store them as attachments.  The attachment is called `Event.ics`.

* FIX:  Only scroll vertically

Fixes #135

* Fix the URL parser to only parse URLs

Mike raised a valid issue where the parser was picking up the first word of each paragraph as a match.  This should resolve the issue.

1. Modified it to handle URLs without a scheme. (www.google.com)
2. Added ability to handle `mailto` and `tel`
3. Added support for punycode URLs.

Fixes #147

* Added the comments that were removed.