Make Feature Usage unlimited
The *snooze&* feature was subscription specifit and was checking for the usage count against the quota.  In the new avatar, NML should have everything unlimited if we are not using any cloud based services.

In case of Cloud based services, we can limit them at some point but I am not sure if we intend to do that.